+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. This projec
 ## [Unreleased]
 ### Added
 - Added JSDoc comments to all functions and important objects. This was then used with the TypeDoc module to generate a documentation site for the package.
+- Added synchronous versions of hash and verify. `hashPasswordSync()` and `verifyHashSync()`.
 
 ## [0.1.3] - 2018-09-06
 ### Added